requirements.txt
flask
line-bot-sdk
openai
gunicorn
redis
